+++ app/versus/page.tsx
@@ -1,587 +1,88 @@
-import { PhotoIcon, UserCircleIcon } from "@heroicons/react/24/solid";
-import { useState } from "react";
+"use client";
+
+import { zodResolver } from "@hookform/resolvers/zod";
+import { resolve } from "path";
+import React from "react";
+import { useForm, SubmitHandler, FieldValues } from "react-hook-form";
+import { object, z } from "zod";
+
+const createImageSchema = z.object({
+  imgName: z.string().max(10, "name can't exceed 10 characters"),
+  object: z.string(),
+  difficulty: z.string(),
+});
+// }).refine(data => data.imgName in imgNames, {
+// message: "Name already taken",
+// path: ["imgName"]
+// } )
+
+
 
 export default function Page() {
+  const {
+    register,
+    handleSubmit,
+    formState: { errors, isSubmitting },
+    getValues,
+  } = useForm({
+    resolver: zodResolver(createImageSchema),
+  });
 
-  return (
+  async function handleFormData(data: FieldValues) {
+    console.log("Hello");
+    await new Promise((resolve) => setTimeout(resolve, 2000));
 
-    <>
-        {/* <FormPartOne submitStatus={oneSubmitted} /> */}
-        <FormPartOne />
-    
-    </>
-  )
-}
+    console.log(data);
+    console.log(data.object);
+  }
 
-function Example() {
-  return (
-    <form>
-      {/* div for space between elements, saves a lot of spacing work */}
-      <div className="space-y-12">
-        <FormButton />
-
-        <FormButton />
-      </div>
-
-      <FormButton />
-    </form>
-  );
-}
-
-// function FormPartOne({submitStatus}: {submitStatus: boolean}) {
-function FormPartOne() {
-    const [oneSubmitted, setOneSubmitted] = useState(false)
-
-
-    function handleSubmit() {
-
-    }
-
-  return (
-    <form onSubmit={handleSubmit}>
-      <div className="space-y-12">
-        <div className="border-b border-gray-900/10 pb-12">
-          <h2 className="text-base font-semibold leading-7 text-gray-900">
-            Part 1: Upload your image
-          </h2>
-          {/* <p className="mt-1 text-sm leading-6 text-gray-600">
-            General help text
-          </p> */}
-
-          <div className="mt-10 grid grid-cols-1 gap-x-6 gap-y-8 sm:grid-cols-6">
-            <div className="sm:col-span-4">
-              <label
-                htmlFor="street-address"
-                className="block text-sm font-medium leading-6 text-gray-900"
-              >
-                Title your captcha for internal use or auto generate
-              </label>
-              <div className="mt-2">
-                <input
-                  type="text"
-                  name="street-address"
-                  id="street-address"
-                  autoComplete="street-address"
-                  className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-                />
-              </div>
-            </div>
-
-            <div className="col-span-full">
-              <label
-                htmlFor="cover-photo"
-                className="block text-sm font-medium leading-6 text-gray-900"
-              >
-                Captcha candidate
-              </label>
-              <p className="text-sm leading-6 text-gray-600">
-                Make sure its 1:1
-              </p>
-              <div className="mt-2 flex justify-center rounded-lg border border-dashed border-gray-900/25 px-6 py-10">
-                <div className="text-center">
-                  <PhotoIcon
-                    className="mx-auto h-12 w-12 text-gray-300"
-                    aria-hidden="true"
-                  />
-                  <div className="mt-4 flex text-sm leading-6 text-gray-600">
-                    <label
-                      htmlFor="file-upload"
-                      className="relative cursor-pointer rounded-md bg-white font-semibold text-indigo-600 focus-within:outline-none focus-within:ring-2 focus-within:ring-indigo-600 focus-within:ring-offset-2 hover:text-indigo-500"
-                    >
-                      <span>Upload a file</span>
-                      <input
-                        id="file-upload"
-                        name="file-upload"
-                        type="file"
-                        className="sr-only"
-                      />
-                    </label>
-                    <p className="pl-1">or drag and drop</p>
-                  </div>
-                  <p className="text-xs leading-5 text-gray-600">
-                    PNG or JPG up to 10MB
-                  </p>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-      <FormButton />
-    </form>
-  );
-}
-
-function FormPartTwo() {
-  return (
-    <div className="border-b border-gray-900/10 pb-12">
-      <h2 className="text-base font-semibold leading-7 text-gray-900">
-        Part 2: Pick from dropdown lists
-      </h2>
-
-      <div className="mt-10 grid grid-cols-1 gap-x-6 gap-y-8 sm:grid-cols-6">
-        <div className="col-span-full">
-          <label
-            htmlFor="cover-photo"
-            className="block text-sm font-medium leading-6 text-gray-900"
-          >
-            This was generated by the object detection alg
-          </label>
-          <div className="mt-2 flex justify-center rounded-lg border border-dashed border-gray-900/25 px-6 py-10">
-            <div className="text-center">
-              <PhotoIcon
-                className="mx-auto h-12 w-12 text-gray-300"
-                aria-hidden="true"
-              />
-              <div className="mt-4 flex text-sm leading-6 text-gray-600">
-                <label
-                  htmlFor="file-upload"
-                  className="relative cursor-pointer rounded-md bg-white font-semibold text-indigo-600 focus-within:outline-none focus-within:ring-2 focus-within:ring-indigo-600 focus-within:ring-offset-2 hover:text-indigo-500"
-                >
-                  <span>Upload a file</span>
-                  <input
-                    id="file-upload"
-                    name="file-upload"
-                    type="file"
-                    className="sr-only"
-                  />
-                </label>
-                <p className="pl-1">or drag and drop</p>
-              </div>
-              <p className="text-xs leading-5 text-gray-600">
-                PNG or JPG up to 10MB
-              </p>
-            </div>
-          </div>
-        </div>
-
-        <div className="sm:col-span-3">
-          <label
-            htmlFor="country"
-            className="block text-sm font-medium leading-6 text-gray-900"
-          >
-            Country
-          </label>
-          <div className="mt-2">
-            <select
-              id="country"
-              name="country"
-              autoComplete="country-name"
-              className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:max-w-xs sm:text-sm sm:leading-6"
-            >
-              <option>United States</option>
-              <option>Canada</option>
-              <option>Mexico</option>
-            </select>
-          </div>
-        </div>
-
-        <br></br>
-        <div className="sm:col-span-3">
-          <label
-            htmlFor="country"
-            className="block text-sm font-medium leading-6 text-gray-900"
-          >
-            Country
-          </label>
-          <div className="mt-2">
-            <select
-              id="country"
-              name="country"
-              autoComplete="country-name"
-              className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:max-w-xs sm:text-sm sm:leading-6"
-            >
-              <option>United States</option>
-              <option>Canada</option>
-              <option>Mexico</option>
-            </select>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function FormPartThree() {
-  return (
-    <div className="border-b border-gray-900/10 pb-12">
-      <h2 className="text-base font-semibold leading-7 text-gray-900">
-        Part 3: Review it
-      </h2>
-
-      <div className="mt-10 grid grid-cols-1 gap-x-6 gap-y-8 sm:grid-cols-6">
-        <div className="col-span-full">
-          <label
-            htmlFor="cover-photo"
-            className="block text-sm font-medium leading-6 text-gray-900"
-          >
-            This was generated by my amazing? code
-          </label>
-          <div className="mt-2 flex justify-center rounded-lg border border-dashed border-gray-900/25 px-6 py-10">
-            <div className="text-center">
-              <PhotoIcon
-                className="mx-auto h-12 w-12 text-gray-300"
-                aria-hidden="true"
-              />
-              <div className="mt-4 flex text-sm leading-6 text-gray-600">
-                <label
-                  htmlFor="file-upload"
-                  className="relative cursor-pointer rounded-md bg-white font-semibold text-indigo-600 focus-within:outline-none focus-within:ring-2 focus-within:ring-indigo-600 focus-within:ring-offset-2 hover:text-indigo-500"
-                >
-                  <span>Upload a file</span>
-                  <input
-                    id="file-upload"
-                    name="file-upload"
-                    type="file"
-                    className="sr-only"
-                  />
-                </label>
-                <p className="pl-1">or drag and drop</p>
-              </div>
-              <p className="text-xs leading-5 text-gray-600">
-                PNG or JPG up to 10MB
-              </p>
-            </div>
-          </div>
-
-          <div className="col-span-full">
-            <label
-              htmlFor="about"
-              className="block text-sm font-medium leading-6 text-gray-900"
-            >
-              Something wrong? Write here (collapse)
-            </label>
-            <div className="mt-2">
-              <textarea
-                id="about"
-                name="about"
-                rows={3}
-                className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-                defaultValue={""}
-              />
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function FormButton() {
-  return (
-    <div className="mt-6 flex items-center justify-end gap-x-6">
-      <button
-        type="button"
-        className="text-sm font-semibold leading-6 text-gray-900"
-      >
-        Reset
-      </button>
-      <button
-        type="submit"
-        className="rounded-md bg-indigo-600 px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
-      >
-        Continue
-      </button>
-    </div>
-  );
-}
-
-function ExtraFormFields() {
   return (
     <>
-      <div className="border-b border-gray-900/10 pb-12">
-        <h2 className="text-base font-semibold leading-7 text-gray-900">
-          Part 2: Pick from dropdown lists
-        </h2>
-
-        <div className="mt-10 grid grid-cols-1 gap-x-6 gap-y-8 sm:grid-cols-6">
-          <div className="col-span-full">
-            <label
-              htmlFor="street-address"
-              className="block text-sm font-medium leading-6 text-gray-900"
-            >
-              Street address
-            </label>
-            <div className="mt-2">
-              <input
-                type="text"
-                name="street-address"
-                id="street-address"
-                autoComplete="street-address"
-                className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-              />
-            </div>
-          </div>
-
-          <div className="sm:col-span-2 sm:col-start-1">
-            <label
-              htmlFor="city"
-              className="block text-sm font-medium leading-6 text-gray-900"
-            >
-              City
-            </label>
-            <div className="mt-2">
-              <input
-                type="text"
-                name="city"
-                id="city"
-                autoComplete="address-level2"
-                className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-              />
-            </div>
-          </div>
-
-          <div className="sm:col-span-2">
-            <label
-              htmlFor="region"
-              className="block text-sm font-medium leading-6 text-gray-900"
-            >
-              State / Province
-            </label>
-            <div className="mt-2">
-              <input
-                type="text"
-                name="region"
-                id="region"
-                autoComplete="address-level1"
-                className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-              />
-            </div>
-          </div>
-
-          <div className="sm:col-span-2">
-            <label
-              htmlFor="postal-code"
-              className="block text-sm font-medium leading-6 text-gray-900"
-            >
-              ZIP / Postal code
-            </label>
-            <div className="mt-2">
-              <input
-                type="text"
-                name="postal-code"
-                id="postal-code"
-                autoComplete="postal-code"
-                className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-              />
-            </div>
-          </div>
-        </div>
-        <div className="sm:col-span-4">
-          <label
-            htmlFor="username"
-            className="block text-sm font-medium leading-6 text-gray-900"
-          >
-            Username
-          </label>
-          <div className="mt-2">
-            <div className="flex rounded-md shadow-sm ring-1 ring-inset ring-gray-300 focus-within:ring-2 focus-within:ring-inset focus-within:ring-indigo-600 sm:max-w-md">
-              <span className="flex select-none items-center pl-3 text-gray-500 sm:text-sm">
-                workcation.com/
-              </span>
-              <input
-                type="text"
-                name="username"
-                id="username"
-                autoComplete="username"
-                className="block flex-1 border-0 bg-transparent py-1.5 pl-1 text-gray-900 placeholder:text-gray-400 focus:ring-0 sm:text-sm sm:leading-6"
-                placeholder="janesmith"
-              />
-            </div>
-          </div>
-        </div>
-
-        <div className="col-span-full">
-          <label
-            htmlFor="about"
-            className="block text-sm font-medium leading-6 text-gray-900"
-          >
-            About
-          </label>
-          <div className="mt-2">
-            <textarea
-              id="about"
-              name="about"
-              rows={3}
-              className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-              defaultValue={""}
+      <form onSubmit={handleSubmit(handleFormData)}>
+        <ul>
+          <li>
+            <label htmlFor="imgName">Image Name:</label>
+            <input
+              {...register("imgName")}
+              type="text"
+              name="imgName"
+              className="border-2"
+              placeholder="potat"
             />
-          </div>
-          <p className="mt-3 text-sm leading-6 text-gray-600">
-            Write a few sentences about yourself.
-          </p>
-        </div>
+          </li>
+          <li>
+            {errors.imgName && (
+              <p className="text-red-500">{`${errors.imgName.message}`}</p>
+            )}
+          </li>
+          <li>
+            <label htmlFor="object">object:</label>
+            <input
+              {...register("object")}
+              type="text"
+              name="object"
+              className="border-2"
+            />
+          </li>
+          <li>
+            <label htmlFor="difficulty">Difficulty:</label>
+            <input
+              {...register("difficulty")}
+              type="text"
+              name="difficulty"
+              className="border-2"
+            />
+          </li>
+        </ul>
 
-        <div className="mt-10 grid grid-cols-1 gap-x-6 gap-y-8 sm:grid-cols-6">
-          <div className="sm:col-span-3">
-            <label
-              htmlFor="first-name"
-              className="block text-sm font-medium leading-6 text-gray-900"
-            >
-              First name
-            </label>
-            <div className="mt-2">
-              <input
-                type="text"
-                name="first-name"
-                id="first-name"
-                autoComplete="given-name"
-                className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-              />
-            </div>
-          </div>
-
-          <div className="sm:col-span-3">
-            <label
-              htmlFor="last-name"
-              className="block text-sm font-medium leading-6 text-gray-900"
-            >
-              Last name
-            </label>
-            <div className="mt-2">
-              <input
-                type="text"
-                name="last-name"
-                id="last-name"
-                autoComplete="family-name"
-                className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-              />
-            </div>
-          </div>
-        </div>
-      </div>
-
-      <div className="border-b border-gray-900/10 pb-12">
-        <h2 className="text-base font-semibold leading-7 text-gray-900">
-          Notifications
-        </h2>
-        <p className="mt-1 text-sm leading-6 text-gray-600">
-          We will always let you know about important changes, but you pick what
-          else you want to hear about.
-        </p>
-
-        <div className="mt-10 space-y-10">
-          <fieldset>
-            <legend className="text-sm font-semibold leading-6 text-gray-900">
-              By Email
-            </legend>
-            <div className="mt-6 space-y-6">
-              <div className="relative flex gap-x-3">
-                <div className="flex h-6 items-center">
-                  <input
-                    id="comments"
-                    name="comments"
-                    type="checkbox"
-                    className="h-4 w-4 rounded border-gray-300 text-indigo-600 focus:ring-indigo-600"
-                  />
-                </div>
-                <div className="text-sm leading-6">
-                  <label
-                    htmlFor="comments"
-                    className="font-medium text-gray-900"
-                  >
-                    Comments
-                  </label>
-                  <p className="text-gray-500">
-                    Get notified when someones posts a comment on a posting.
-                  </p>
-                </div>
-              </div>
-              <div className="relative flex gap-x-3">
-                <div className="flex h-6 items-center">
-                  <input
-                    id="candidates"
-                    name="candidates"
-                    type="checkbox"
-                    className="h-4 w-4 rounded border-gray-300 text-indigo-600 focus:ring-indigo-600"
-                  />
-                </div>
-                <div className="text-sm leading-6">
-                  <label
-                    htmlFor="candidates"
-                    className="font-medium text-gray-900"
-                  >
-                    Candidates
-                  </label>
-                  <p className="text-gray-500">
-                    Get notified when a candidate applies for a job.
-                  </p>
-                </div>
-              </div>
-              <div className="relative flex gap-x-3">
-                <div className="flex h-6 items-center">
-                  <input
-                    id="offers"
-                    name="offers"
-                    type="checkbox"
-                    className="h-4 w-4 rounded border-gray-300 text-indigo-600 focus:ring-indigo-600"
-                  />
-                </div>
-                <div className="text-sm leading-6">
-                  <label htmlFor="offers" className="font-medium text-gray-900">
-                    Offers
-                  </label>
-                  <p className="text-gray-500">
-                    Get notified when a candidate accepts or rejects an offer.
-                  </p>
-                </div>
-              </div>
-            </div>
-          </fieldset>
-          <fieldset>
-            <legend className="text-sm font-semibold leading-6 text-gray-900">
-              Push Notifications
-            </legend>
-            <p className="mt-1 text-sm leading-6 text-gray-600">
-              These are delivered via SMS to your mobile phone.
-            </p>
-            <div className="mt-6 space-y-6">
-              <div className="flex items-center gap-x-3">
-                <input
-                  id="push-everything"
-                  name="push-notifications"
-                  type="radio"
-                  className="h-4 w-4 border-gray-300 text-indigo-600 focus:ring-indigo-600"
-                />
-                <label
-                  htmlFor="push-everything"
-                  className="block text-sm font-medium leading-6 text-gray-900"
-                >
-                  Everything
-                </label>
-              </div>
-              <div className="flex items-center gap-x-3">
-                <input
-                  id="push-email"
-                  name="push-notifications"
-                  type="radio"
-                  className="h-4 w-4 border-gray-300 text-indigo-600 focus:ring-indigo-600"
-                />
-                <label
-                  htmlFor="push-email"
-                  className="block text-sm font-medium leading-6 text-gray-900"
-                >
-                  Same as email
-                </label>
-              </div>
-              <div className="flex items-center gap-x-3">
-                <input
-                  id="push-nothing"
-                  name="push-notifications"
-                  type="radio"
-                  className="h-4 w-4 border-gray-300 text-indigo-600 focus:ring-indigo-600"
-                />
-                <label
-                  htmlFor="push-nothing"
-                  className="block text-sm font-medium leading-6 text-gray-900"
-                >
-                  No push notifications
-                </label>
-              </div>
-            </div>
-          </fieldset>
-        </div>
-      </div>
+        <button
+          type="submit"
+          disabled={isSubmitting}
+          className="bg-blue-400 disabled:bg-slate-500"
+        >
+          Submit
+        </button>
+      </form>
     </>
   );
 }
