+++ package.json
@@ -12,6 +12,7 @@
     "@headlessui/react": "^1.7.19",
     "@heroicons/react": "^2.1.3",
     "@hookform/resolvers": "^3.4.2",
+    "@uploadcare/react-uploader": "^0.5.3",
     "@vercel/postgres": "^0.8.0",
     "drizzle-orm": "^0.30.10",
     "next": "14.2.3",
